+++ CHANGELOG.md
@@ -3,4 +3,4 @@
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
 ## 1.0.0
-- Setup installable custom cursor
+- Setup installable super cursor